Added code to enforce uniqueness of note identifiers in Decent Notes.
Using the ReplicatingArray does not guarantee uniqueness of ids. Imagine, for example, that on one device we move an item to a different location, and also move it on the second device, but to somewhere else. This will result in the same deletion on each device, but two different insertions, giving duplicate entries.